use microprofile.LRA instead of mpLRA for packages
