+++ dev/io.openliberty.microprofile.LRA.1.0.internal/.classpath
