+++ dev/io.openliberty.microprofile.LRA.1.0.internal/.settings/bndtools.core.prefs
