+++ dev/io.openliberty.microprofile.LRA.1.0.internal/.settings/org.eclipse.core.resources.prefs
