+++ dev/io.openliberty.microprofile.LRA.1.0.internal/resources/OSGI-INF/l10n/metatype.properties
