+++ dev/io.openliberty.microprofile.LRA.1.0.internal/resources/io/openliberty/microprofile/LRA/internal/resources/Lra.nlsprops
