+++ dev/io.openliberty.microprofile.LRA.1.0.internal/src/io/openliberty/microprofile/LRA/internal/LraConfig.java
@@ -8,7 +8,7 @@
  * Contributors:
  *     IBM Corporation - initial API and implementation
  *******************************************************************************/
-package io.openliberty.mpLRA.internal;
+package io.openliberty.microprofile.LRA.internal;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -26,7 +26,7 @@ import com.ibm.websphere.ras.TraceComponent;
 /**
  *
  */
-@Component(name = "com.ibm.ws.lra.config", service = LraConfig.class, configurationPid = "com.ibm.ws.lra.config", configurationPolicy = ConfigurationPolicy.REQUIRE, immediate = true, property = { "service.vendor=IBM" })
+@Component(name = "io.openliberty.microprofile.LRA.1.0.internal.config", service = LraConfig.class, configurationPid = "io.openliberty.microprofile.LRA.1.0.internal.config", configurationPolicy = ConfigurationPolicy.REQUIRE, immediate = true, property = { "service.vendor=IBM" })
 public class LraConfig {
     private static final TraceComponent tc = Tr.register(LraConfig.class);
