+++ dev/io.openliberty.microprofile.LRA.1.0.internal/src/io/openliberty/microprofile/LRA/internal/LraFilterComponent.java
@@ -8,7 +8,7 @@
  * Contributors:
  *     IBM Corporation - initial API and implementation
  *******************************************************************************/
-package io.openliberty.mpLRA.internal;
+package io.openliberty.microprofile.LRA.internal;
 
 import java.net.URI;
 import java.net.URISyntaxException;
